+++ src/lib/components/component-collection.ts
@@ -65,6 +65,13 @@ import { TerraFormContainerWrapperComponent } from './forms/form/form-container/
 import { TableSettingsComponent } from './table/table-settings/table-settings.component';
 import { TableSettingsDialogComponent } from './table/table-settings/dialog/table-settings-dialog.component';
 import { MultiSelectComponent } from './forms/form/form-components/multi-select/multi-select.component';
+import { ColorPickerComponent } from './forms/form/form-components/color-picker/color-picker.component';
+import { DatePickerComponent } from './forms/form/form-components/datepicker/datepicker.component';
+import { TextInputComponent } from './forms/form/form-components/text-input/text-input.component';
+import { CheckboxComponent } from './forms/form/form-components/checkbox/checkbox.component';
+import { SelectComponent } from './forms/form/form-components/select/select.component';
+import { NumberInputComponent } from './forms/form/form-components/number-input/number-input.component';
+import { TextAreaComponent } from './forms/form/form-components/text-area/text-area.component';
 
 export const exportedComponents: Array<Type<any>> = [
     TerraAlertPanelComponent,
@@ -136,5 +143,12 @@ export const components: Array<Type<any>> = [
     TerraBaseEditorComponent,
     TableSettingsDialogComponent,
     TerraFormContainerWrapperComponent,
-    MultiSelectComponent
+    MultiSelectComponent,
+    ColorPickerComponent,
+    DatePickerComponent,
+    TextInputComponent,
+    CheckboxComponent,
+    SelectComponent,
+    NumberInputComponent,
+    TextAreaComponent
 ];