+++ src/lib/components/forms/form/form-components/datepicker/datepicker.interface.ts
@@ -1,0 +1,5 @@
+import { TerraFormComponentBaseInterface } from '../terra-form-component-base.interface';
+
+export interface DatepickerInterface extends TerraFormComponentBaseInterface {
+    displayDateFormat: string;
+}