+++ src/lib/terra-components.module.ts
@@ -19,6 +19,9 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSelectModule } from '@angular/material/select';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { A11yModule } from '@angular/cdk/a11y';
+import { MatInputModule } from '@angular/material/input';
 
 @NgModule({
     imports: [
@@ -39,7 +42,11 @@ import { MatSelectModule } from '@angular/material/select';
         MatIconModule,
         MatCheckboxModule,
         MatFormFieldModule,
-        MatSelectModule
+        MatDatepickerModule,
+        A11yModule,
+        MatFormFieldModule,
+        MatSelectModule,
+        MatInputModule
     ],
     declarations: [...components, ...directives, ...pipes],
     exports: [...exportedComponents, ...exportedDirectives, ...pipes]